# swagger.yaml
swagger: '2.0'
info:
  description: Automatically produces natural language news reports from EuroStat data.
  title: EuroStat news NLG API
  version: 2.0.0
paths:
  /health:
    options:
      description: "Returns a JSON array of language codes supported by the /eunlg endpoint."
      responses:
        '200':
          description: OK
    get:
      description: "Returns the version number of the system as well. A reply indicates the system is healthy."
      produces:
        - application/json
      responses:
        '200':
          description: OK
          schema:
            type: object
            required:
              - version
            properties:
              version:
                type: string
                example: "0.1.0"

  /languages:
    options:
      description: "Describes the available HTTP methods for this end point."
      responses:
        '200':
          description: OK
    get:
      description: "Returns a JSON array of language codes supported by the /eunlg endpoint."
      produces:
        - application/json
      responses:
        '200':
          description: OK
          schema:
            type: object
            required:
              - languages
            properties:
              languages:
                type: array
                items:
                  type: string
                example:
                  - en
                  - fi
  /datasets:
    options:
      description: "Describes the available HTTP methods for this end point."
      responses:
        '200':
          description: OK
    post:
      description: >-
        Returns a JSON array of datasets supported by the /eunlg and /location
        endpoints.
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: language
          description: An object containing any language identifier returned by /languages
          schema:
            type: object
            required:
              - language
            properties:
              language:
                type: string
                example: "en"
      responses:
        '200':
          description: OK
          schema:
            type: object
            required:
              - datasets
            properties:
              datasets:
                type: array
                items:
                  type: string
                example:
                  - cphi
                  - env
        '400':
          description: Error
          schema:
            type: object
            properties:
              error:
                type: string
                example:
                  "Invalid value for 'language', query /languages for valid options."
  /locations:
    options:
      description: "Describes the available HTTP methods for this end point."
      responses:
        '200':
          description: OK
    post:
      description: >-
        Query for valid locations supported by the /eunlg endpoint for a certain
        dataset
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: dataset
          description: An object containing any identifier returned by /datasets
          schema:
            type: object
            required:
              - dataset
            properties:
              dataset:
                type: string
                example: "cphi"
      responses:
        '200':
          description: OK
          schema:
            type: object
            properties:
              locations:
                type: array
                items:
                  type: string
                example:
                  - AT
                  - ET
        '400':
          description: Error
          schema:
            type: object
            properties:
              error:
                type: string
                example:
                  "Invalid value for 'dataset', query /datasets for valid options."
  /eunlg:
    options:
      description: "Describes the available HTTP methods for this end point."
      responses:
        '200':
          description: OK
    post:
      description: Generate a report
      produces:
        - application/json
      consumes:
        - application/json
      parameters:
        - in: body
          name: query
          description: An object defining what dataset to query from, focused on which location, and what language to output.
          schema:
            type: object
            required:
              - location
              - dataset
            properties:
              dataset:
                type: string
              location:
                type: string
              language:
                type: string
      responses:
        '200':
          description: OK
          schema:
            type: object
            properties:
              location:
                type: string
                example: FI
              location_type:
                type: string
                example: C
              language:
                type: string
                example: en
              header:
                type: string
                example: <h1>...</h1>
              body:
                type: string
                example: <p>...</p>
        '400':
          description: Missing or invalid inputs
          schema:
            type: object
            properties:
              error:
                type: string
                example:
                  "Invalid value for 'location', query /locations for valid options."